requirements.txt
mysqlclient
sqlalchemy
uvicorn
fastapi
alembic
python-dotenv
sqlalchemy_utils
PyJWT